[Tooltip] Implement support for Tooltip.Placement and PlacementTarget
Add support for the new SL3 features for placing tooltips relative to
existing UIElements. When a tooltip is manually opened we need to use
the regular Tooltip properties, but when it is auto-opened by the
TooltipService we need to use the TooltipService equivalents to override
the locally set values (if any). Makes 672 a bit better.