[ListBox] Handle scrolling to virtualized elements properly.
If the listbox is backed by a VirtualizingPanel we need to scroll to
the index of the virtualized element instead of a pixel offset. This
is because only a subset of the elements are rendered when virtualizing
and the pixel offset is useless.

We also need to correctly detect the orientation of the panel when a
VirtualizingStackPanel is used.